src/data/data_collector.py
import requests
from bs4 import BeautifulSoup
import pandas as pd
import aiohttp
import asyncio
from typing import Dict, List, Any
import logging
import os

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class DataCollector:
    def __init__(self):
        self.base_url = "https://api.football-data.org/v2"
        self.api_key = os.getenv("FOOTBALL_DATA_API_KEY")
        self.headers = {"X-Auth-Token": self.api_key}
        
    async def get_team_stats(self, team_name: str) -> Dict[str, Any]:
        """Bir takımın son maçlarındaki istatistiklerini toplar."""
        try:
            stats = {
                'recent_matches': await self._get_recent_matches(team_name),
                'head_to_head': await self._get_head_to_head(team_name),
                'season_stats': await self._get_season_stats(team_name),
                'player_stats': await self._get_player_stats(team_name)
            }
            return stats
        except Exception as e:
            logger.error(f"Error collecting stats for {team_name}: {str(e)}")
            raise
    
    async def _get_recent_matches(self, team_name: str, limit: int = 10) -> List[Dict]:
        """Son maçların detaylarını getirir."""
        async with aiohttp.ClientSession() as session:
            async with session.get(
                f"{self.base_url}/teams",
                headers=self.headers,
                params={"name": team_name}
            ) as response:
                team_data = await response.json()
                team_id = team_data['teams'][0]['id']
                
                async with session.get(
                    f"{self.base_url}/teams/{team_id}/matches",
                    headers=self.headers,
                    params={"limit": limit, "status": "FINISHED"}
                ) as matches_response:
                    matches = await matches_response.json()
                    return self._process_matches(matches['matches'])
    
    async def _get_head_to_head(self, team1: str, team2: str) -> List[Dict]:
        """İki takım arasındaki geçmiş maçları getirir."""
        # API'den head-to-head verilerini al
        pass
    
    async def _get_season_stats(self, team_name: str) -> Dict:
        """Sezon istatistiklerini getirir."""
        # Sezon istatistiklerini topla
        pass
    
    async def _get_player_stats(self, team_name: str) -> List[Dict]:
        """Takım oyuncularının istatistiklerini getirir."""
        # Oyuncu istatistiklerini topla
        pass
    
    def _process_matches(self, matches: List[Dict]) -> List[Dict]:
        """Ham maç verilerini işler ve analiz için hazırlar."""
        processed_matches = []
        for match in matches:
            processed_match = {
                'date': match['utcDate'],
                'home_team': match['homeTeam']['name'],
                'away_team': match['awayTeam']['name'],
                'home_score': match['score']['fullTime']['homeTeam'],
                'away_score': match['score']['fullTime']['awayTeam'],
                'winner': match['score']['winner'],
                'status': match['status']
            }
            processed_matches.append(processed_match)
        return processed_matches
    
    def get_odds_data(self, match_id: str) -> Dict:
        """Maç için bahis oranlarını getirir."""
        # Bahis oranlarını topla
        pass 