requirements.txt
flask==2.0.1
tensorflow==2.19.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
requests>=2.25.0
python-dotenv>=0.19.0
waitress>=2.0.0
joblib>=1.0.0
Werkzeug==2.0.1
beautifulsoup4==4.9.3
aiohttp==3.8.1 